+++ lib/oh_my_flutter.dart
@@ -19,6 +19,8 @@ export 'src/telephony.dart' show Telephony;
 export 'src/whatsapp.dart' show Whatsapp;
 export 'src/widgets/controlled_visibility/controlled_visibility.dart'
     show ControlledVisibility, ControlledVisibilityController;
+export 'src/widgets/marquee/marquee.dart' show Marquee;
+export 'src/widgets/marquee/marquee_direction.dart' show MarqueeDirection;
 export 'src/widgets/motion/motion.dart'
     show
         FadeInMotionEffect,
@@ -27,7 +29,8 @@ export 'src/widgets/motion/motion.dart'
         MotionEffect,
         MotionPlayback,
         MoveMotionEffect,
-        ScaleInMotionEffect;
+        ScaleInMotionEffect,
+        TextMotion;
 export 'src/widgets/pause_animations/pause_animations.dart' show PauseAnimations;
 export 'src/widgets/route_settled/route_settled.dart' show RouteSettled;
 export 'src/widgets/sequence/sequence.dart' show Sequence, SequenceController, SequenceTransitionBuilder;